Overall website enhancement (#2)
# Summary
Without moving too much around, as to not disturb the current workflow
of the developers, I managed to fix some of the visual errors in both
the normal and responsive view along with some errors in the main.js
file. I also implemented prettier as a developer dependency as to keep
all the current developers in the same page with the formatting of the
codebase.

# What's exactly "new"?
- Fixed an error in the main.js file with some '$' characters even
though there was no JQuery being used or referenced.
- Applied better spacing between the chatbot div and the navbar along
with a better spacing and center of the buttons by adding the class
"buttons-container" and adding some scss.
- Added the functionallity for the navbar to show and not show once the
user enters the mobile menu
- Added .prettierignore file as to not accidentally **format the
entirety of node_modules and such**. Better explanation at [this
commit's
comment](https://github.com/AngelloFD/InterChatWebsite/commit/5b16fef5e4497f90d8c52ad9baed9041256545b1)
- Changed .gitignore file as to ignore folders that shouldn't have been
committed. Complete explanation at [the same commit's
comment](https://github.com/AngelloFD/InterChatWebsite/commit/5b16fef5e4497f90d8c52ad9baed9041256545b1)
- Added script to automatically format all the codebase given
.prettierignore file and the .editorconfig file.

# What did you not do?
- I did not completed the referencing to the other pages as they seemed
incomplete as my vision and design for said pages could clash with the
developer's own.
- I did not added some form of organization of the files as, in my
experience, Github Pages is somewhat inconsistent with the referencing
of such.
- I did not upgraded or touched any of the existing dependencies as it
could lead to broken features and a much more cautious revision should
be done in that regard.

I'll be more than happy to fix any or move anything in regards of the
features or fixes I made, if the developers deem necessary.
Adding my grain of sand to the project since I am very interested in the
theme of it.

---------

Co-authored-by: dev-737 <73829355+dev-737@users.noreply.github.com>